Timeline CR (#413)
* Timeline CR

* Timeline CR

* Timeline CR add edit test

* fix yml fixture issue

* style

* fix tests

* fix tests

* fix tests

* fix tests

* fix tests

* fix tests

* ItemList improvement

* ItemList improvement

* ItemList improvement

* fix tests

* fix tests

* fix tests

* fix tests

* fix tests

* fix tests

* update tests

* update tests

* update tests

* update tests

* use Translations for monthNames

* Update app/javascript/components/Line/components/Line.jsx


* Update app/helpers/items_helper.rb


* CR

* revert locales changes

* CR

* CR

* CR

* CR

+++ app/models/field/date_time.rb
@@ -125,12 +125,12 @@ class Field::DateTime < ::Field
   end
 
   def order_items_by(direction: 'ASC')
-    "NULLIF(items.data->'#{uuid}'->>'Y', '')::int #{direction},
-    NULLIF(items.data->'#{uuid}'->>'M', '')::int #{direction},
-    NULLIF(items.data->'#{uuid}'->>'D', '')::int #{direction},
-    NULLIF(items.data->'#{uuid}'->>'h', '')::int #{direction},
-    NULLIF(items.data->'#{uuid}'->>'m', '')::int #{direction},
-    NULLIF(items.data->'#{uuid}'->>'s', '')::int #{direction}"
+    "NULLIF(items.data->'#{uuid}'->>'Y', '')::bigint #{direction},
+     NULLIF(items.data->'#{uuid}'->>'M', '')::bigint #{direction},
+     NULLIF(items.data->'#{uuid}'->>'D', '')::bigint #{direction},
+     NULLIF(items.data->'#{uuid}'->>'h', '')::bigint #{direction},
+     NULLIF(items.data->'#{uuid}'->>'m', '')::bigint #{direction},
+     NULLIF(items.data->'#{uuid}'->>'s', '')::bigint #{direction}"
   end
 
   def search_conditions_as_hash(locale)
